requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
plotly>=5.22
pydeck>=0.9.1
pydantic>=2.8
python-dateutil>=2.9
xlsxwriter>=3.2
openpyxl>=3.1
